+++ editor/instancer/voxel_instancer_editor_plugin.h
@@ -6,6 +6,7 @@
 namespace zylann::voxel {
 
 class VoxelInstancer;
+class VoxelInstancerStatView;
 
 class VoxelInstancerEditorPlugin : public EditorPlugin {
 	GDCLASS(VoxelInstancerEditorPlugin, EditorPlugin)
@@ -17,7 +18,12 @@ public:
 	void make_visible(bool visible) override;
 
 private:
+	bool toggle_stat_view();
+	void _on_menu_item_selected(int id);
+
+	MenuButton *_menu_button = nullptr;
 	VoxelInstancer *_node = nullptr;
+	VoxelInstancerStatView *_stat_view = nullptr;
 };
 
 } // namespace zylann::voxel